logic fixes
